numaan/requirements.txt
torch>=2.1
torchvision>=0.16
facenet-pytorch>=2.6.0
opencv-python>=4.8
numpy>=1.24
pillow>=10.0
tqdm>=4.66
pyyaml>=6.0
scikit-learn>=1.3
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.9
onnxruntime>=1.17